LayoutTests/inspector/styles/protocol-css-regions-commands.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>

function createDynamicElements()
{
    var frameDoc = window.frames[0].document;

    var el = frameDoc.createElement("article");
    el.style.webkitFlowInto = "flow4";
    frameDoc.body.appendChild(el);

    el = frameDoc.createElement("div");
    el.style.webkitFlowFrom = "flow4";
    frameDoc.body.appendChild(el);

    runTest();
}

function createNamedFlow()
{
    var article = document.createElement("article");
    article.id = "tmpArticle";
    article.style.webkitFlowInto = "tmpNamedFlow";

    document.body.appendChild(article);
}

function removeNamedFlow()
{
    var article = document.getElementById("tmpArticle");

    document.body.removeChild(article);
}

function test()
{
    WebInspector.showPanel("elements");
    InspectorTest.runTestSuite([
        function testGetNamedFlowCollection1(next)
        {
            WebInspector.domAgent.requestDocument(documentCallback);

            function documentCallback(document)
            {
                WebInspector.cssModel.getNamedFlowCollectionAsync(document.id, namedFlowCallback);
            }

            function namedFlowCallback(namedFlows)
            {
                InspectorTest.addResult("=== CSS Named Flows in main document ===");

                if (!namedFlows) {
                    InspectorTest.addResult("[!] Failed to get Named Flows");
                    InspectorTest.completeTest();
                    return;
                }

                namedFlows.sort(function (nf1, nf2) {
                    return ((nf1.name == nf2.name) ? 0 : ((nf1.name > nf2.name) ? 1 : -1));
                });

                for (var i = 0; i < namedFlows.length; ++i)
                    printNamedFlow(namedFlows[i]);

                next();
            }
        },

        function testGetNamedFlowCollection2(next)
        {
            WebInspector.domAgent.requestDocument(documentCallback);

            function documentCallback(document)
            {
                WebInspector.domAgent.querySelector(document.id, "#fake-document", querySelectorCallback);
            }

            function querySelectorCallback(nodeId)
            {
                WebInspector.cssModel.getNamedFlowCollectionAsync(nodeId, namedFlowCallback);
            }

            function namedFlowCallback(namedFlows)
            {
                InspectorTest.addResult("=== CSS Named Flows in #fake-document ===");

                if (namedFlows) {
                    InspectorTest.addResult("[!] Failed");
                    InspectorTest.completeTest();
                    return;
                }

                InspectorTest.addResult("#fake-document is not a document")
                next();
            }
        },

        function testGetFlowByName1(next)
        {
            WebInspector.domAgent.requestDocument(documentCallback);

            function documentCallback(document)
            {
                WebInspector.cssModel.getFlowByNameAsync(document.id, "flow2", namedFlowCallback);
            }

            function namedFlowCallback(namedFlow)
            {
                InspectorTest.addResult("=== Named Flow \"flow2\" from main document ===");

                if (!namedFlow) {
                    InspectorTest.addResult("[!] Failed to get Named Flow");
                    InspectorTest.completeTest();
                    return;
                }

                printNamedFlow(namedFlow);

                next();
            }
        },

        function testGetFlowByName2(next)
        {
            WebInspector.domAgent.requestDocument(documentCallback);

            function documentCallback(document)
            {
                WebInspector.cssModel.getFlowByNameAsync(document.id, "flow4", namedFlowCallback);
            }

            function namedFlowCallback(namedFlow)
            {
                InspectorTest.addResult("=== Name Flow \"flow4\" from main document ===");

                if (namedFlow) {
                    InspectorTest.addResult("[!] Failed")
                    InspectorTest.completeTest();
                    return;
                }

                InspectorTest.addResult("There is no Named Flow \"flow4\" in the main document");
                next();
            }
        },

        function testNamedFlowCreated(next)
        {
            WebInspector.cssModel.addEventListener(WebInspector.CSSStyleModel.Events.NamedFlowCreated, callback, this);
            InspectorTest.evaluateInPage("createNamedFlow()");

            function callback(event)
            {
                WebInspector.cssModel.removeEventListener(WebInspector.CSSStyleModel.Events.NamedFlowCreated, callback, this);

                if (event.data.name !== "tmpNamedFlow") {
                    Inspector.addResult("[!] Failed");
                    InspectorTest.completeTest();
                    return;
                }

                InspectorTest.addResult("NamedFlowCreated: \"tmpNamedFlow\"");
                next();
            }
        },

        function testNamedFlowRemoved(next)
        {
            WebInspector.cssModel.addEventListener(WebInspector.CSSStyleModel.Events.NamedFlowRemoved, callback, this);
            InspectorTest.evaluateInPage("removeNamedFlow()");

            function callback(event)
            {
                WebInspector.cssModel.removeEventListener(WebInspector.CSSStyleModel.Events.NamedFlowRemoved, callback, this);
                if (event.data.name !== "tmpNamedFlow") {
                    Inspector.addResult("[!] Failed");
                    InspectorTest.completeTest();
                    return;
                }

                InspectorTest.addResult("NamedFlowRemoved: \"tmpNamedFlow\"");
                next();
            }
        }
    ]);

    function printNamedFlow(namedFlow)
    {
        InspectorTest.addResult("* Named Flow \"" + namedFlow.name + "\"");
        InspectorTest.addResult("Content Elements: " + namedFlow.content.length);
        InspectorTest.addResult("Regions: " + namedFlow.regions.length);
    }
}
</script>
</head>

<body onload="createDynamicElements()">
<p>
Tests the following commands and events:
<ul>
    <li>getNamedFlowCollection <a href="https://bugs.webkit.org/show_bug.cgi?id=91607">Bug 91607</a></li>
    <li>getFlowByName <a href="https://bugs.webkit.org/show_bug.cgi?id=91855">Bug 91855</a></li>
    <li>namedFlowCreated <a href="https://bugs.webkit.org/show_bug.cgi?id=92739">Bug 92739</a></li>
    <li>namedFlowRemoved <a href="https://bugs.webkit.org/show_bug.cgi?id=92739">Bug 92739</a></li>
</ul>
</p>

<article style="-webkit-flow-into: flow1"></article>
<div style="-webkit-flow-from: flow1"></div>
<div style="-webkit-flow-from: flow1"></div>
<div style="-webkit-flow-from: flow1"></div>

<article style="-webkit-flow-into: flow2"></article>
<div style="-webkit-flow-from: flow2"></div>
<div style="-webkit-flow-from: flow2"></div>

<article style="-webkit-flow-into: flow3"></article>
<div style="-webkit-flow-from: flow3"></div>

<div id="fake-document"></div>

<iframe></iframe>

</body>
</html>
